locked down posts and comments mutations

+++ src/resolvers/Mutation.js
@@ -108,7 +108,16 @@ const Mutation = {
     );
   },
 
-  async updatePost(parent, args, { prisma }, info) {
+  async updatePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId,
+      },
+    });
+
+    if (!postExists) throw new Error('Unable to update a post!');
     return await prisma.mutation.updatePost(
       {
         data: args.data,
@@ -120,14 +129,15 @@ const Mutation = {
     );
   },
 
-  async createComment(parent, args, { prisma }, info) {
+  async createComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
     return await prisma.mutation.createComment(
       {
         data: {
           text: args.data.text,
           author: {
             connect: {
-              id: args.data.author,
+              id: userId,
             },
           },
           post: {
@@ -141,7 +151,16 @@ const Mutation = {
     );
   },
 
-  async deleteComment(parent, args, { prisma }, info) {
+  async deleteComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId,
+      },
+    });
+
+    if (!commentExists) throw new Error('Unable to delete a comment!');
     return await prisma.mutation.deleteComment(
       {
         where: {
@@ -152,7 +171,16 @@ const Mutation = {
     );
   },
 
-  async updateComment(parent, args, { prisma }, info) {
+  async updateComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId,
+      },
+    });
+
+    if (!commentExists) throw new Error('Unable to update a comment!');
     return await prisma.mutation.updateComment(
       {
         data: args.data,
